[MLIR] Create a chlo.compute_reshape_shape op.
This op handles the broadcasting logic used in tf.Reshape. A single dimension can have a -1 extent and will instead be converted to the number that ensures a valid reshape. This op assumes that such a value exists though.

This CL also removes some restrictions on allowed shapes for tf.Reshape when lowering to HLO as HLO does not need the inputs to have static shapes.

PiperOrigin-RevId: 383251970