Remove unnecessary absl dependencies
These were added in
https://github.com/tensorflow/tensorflow/commit/a39a49009a
and are not necessary. In the mlir-hlo mirror repository, the commit
was https://github.com/tensorflow/mlir-hlo/commit/87424d8e5b, which
didn't rewrite the way that absl is referred to in Google's monorepo
to the way it should be referenced in OSS, so in fact broke the build.

PiperOrigin-RevId: 392780295