Add verifier for mhlo.gather
There is more verification that can be done based on shape inference, but this is a start. I had to delete some tests for highly-specialized lowerings because the input failed verification and I was unable to construct a matching test case. I think those lowerings just actually don't work for that case. Which is fair because the Gather operation is far too complicated. I think a more fruitful avenue would be to have a series of patterns that "canonicalize" (we may not be able to actually make them canonicalizations because this could affect performance on some backends) the Gather to a much simpler form + transpose and reshape operations. It should then be much easier to write patterns against it. One way to do this would be to actually have another op ("SimpleGather"?) that encapsulates this.

PiperOrigin-RevId: 399320417