Lowering of general mhlo.gather to linalg
This is a complete lowering of the gather op and all its weirdnesses.
Accounting for all these makes the lowering pretty fiddly and also
means we miss more efficient representations for the common special
cases. Some of those are already covered by the lowering via
mhlo.torch_index_select, but this is something that should be
revisited. In the meantime this gets us a (hopefully) correct lowering
for the full API.

In addition to the lit tests here, I also spot-checked the output of
running this through IREE against `jax.lax.gather`.

Fixes https://github.com/tensorflow/mlir-hlo/issues/16

PiperOrigin-RevId: 400063077