[MLIR][MHLO] Make broadcast propagation unbounded on the number of iterations
Broadcast propagation will eventually reach a fix point. Depending on the size
of the input IR the number of iterations in the greedy pattern rewrite mechanism
can vary. For general correctness, the number of iterations must be unbounded.
It will typically be very small. Today, the default 10 operations have always
been sufficient.

PiperOrigin-RevId: 401212212