Remove the LHLO to Linalg ConvOp lowering.
The Linalg ConvOp has been replaced by its OpDSL counter parts and is about to disappear. This patch removes the remaining LHLO to Linalg conversion pattern lowering to Linalg ConvOp.

NOTE: The LHLO to Linalg path to the best of our knowledge has no uses anymore. We thus plan to retire it in the near future.
PiperOrigin-RevId: 401220311