PR #52299: Copy removal
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/52299

This copy removal pass is based on the UserangeAnalysis. First, all Operations implementing the CopyOpInterface are found, then the useranges of the source and target of the CopyOpInterface are intersected. The only Operation in the intersection must be the CopyOp or Operations that are within the nested region if the CopyOp implements a Region. If this is the case, we can remove the CopyOp.

This PR includes an addition to the UserangeAnalysis to intersect two useranges.

This is the implementation of the Implementation of a Copy Removal Pass #52029.
Copybara import of the project:

--
f54b1b331dec7d228023b8fa562fb58e9635fdf3 by Thomas Schmeyer <Thomas.Schmeyer@dfki.de>:

Implementation of a Copy Removal Pass.

--
24d320808e28ef9e656df3833fe46e2737100601 by Thomas Schmeyer <Thomas.Schmeyer@dfki.de>:

Addressed reviewers comments and fixed a bug with UserangePositions.

PiperOrigin-RevId: 402527815