[tfrt:jit] Insert a copy when returning a dynamic broadcast
We bufferize dynamic broadcasts into a memref reinterpret cast that yields a
memref with affine map. This clashes with the return type of the function that
doesn't support affine maps. Insert a copy for this special case.

This is still a bit of a hack, but I prefer not to invest too much as a
different representation for dynamic broaddcasts is on the horizon.

PiperOrigin-RevId: 405473962