Add optional attributes to specify operand/result layouts in mhlo::CustomCallOp.
A custom call can also have layout constrains on operands and results which can be specified as optional `operand_layouts` and `result_layouts` attributes. The layout attribute is an array of rank-1 index tensors and the i-th layout attribute specifies the layout for i-th operand/result.

The `operand_layouts` & `result_layouts` attributes can be specified under the following constraints checked in the verifier:
1) Either both `operand_layouts` and `result_layouts` are specified or none.
2) None of the operands are of tuple type.
3) None of the results are of tuple type except the common case of single tuple result packing non-tuple values is allowed. In this case the i-th `result_layouts` attribute specifies the layout of i-th element in the result tuple.

PiperOrigin-RevId: 410823230