[ShapeAnalysis] Rename functions and rephrase comments in the presence of scalar analysis
The shape analysis looks at shapes and values simultaneously and returns a list
of symbolic expressions. For values (e.g., a shape tensor), these correspond to
the individual element values, for shapes (e.g., the shape of a tensor), these
correspond to the individual dimensions.

This CL is an attempt at distinguishing more clearly between the value and the
shape analysis. It also considers the fact that scalar values have become
subject to this analysis.

PiperOrigin-RevId: 412289108