[ShapeAnalysis] Eliminate more redundant shape constraints in real-world cases
Eliminate more redundant `cstr_reshapable` constraints. Collect all the factors
before comparing the number of elements and the dynamic shape for compatibility.
Previously, we did this only for the symbolic expression that falls out of the
`num_elements` op, not for the individual dimensions of the dynamic shape. With
this, we are able to eliminate the `cstr_reshapable` in the real-world
integration test.

PiperOrigin-RevId: 412450559