Add test for parsing of mhlo.reduce compact form when --mlir-print-local-scope=false
When printing without the local scope, the locations are printed at the end of
the file and refered to with aliases. When parsing back we don't have the
location at hand and instead the parser handled itself the remapping of the
location when parsing is done.
This use to crash and has been fixed in MLIR.

Fixes https://github.com/tensorflow/mlir-hlo/issues/25

PiperOrigin-RevId: 422838347