Add a pass to restrict max rank in the mhlo dialect
The pass currently has a pattern to breakdown a sequence of reshape -> transpose -> reshape ops into multiple sequences. The original sequence is used to merge a part of spatial dimensions with the batch dimension so we can restrict the max rank by doing the operation for each of the spatial dims.

This transformation could potentially increases the overhead but helps in reducing dimensionality. This could be useful in backends that don't support higher ranked tensors.

PiperOrigin-RevId: 447535424