Lower mhlo.round_nearest_afz to math.round
The lowering we are currently using is incorrect in cases when the number is close to 0.5. E.g., 0.49999997f in FP32 will get rounded to 1.0f, since 0.49999997f + 0.5f == 1.0f.

There's a dedicated math.round instruction upstream, which rounds correctly, so use that instead.

(We also have the added benefit of removing code duplication.)

PiperOrigin-RevId: 456092038