PR #56844: [mhlo] Add missing dependency to the MhloDialect
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/56844

Recent versions of mlir have moved functionality around, so now the MhloDialect has a dependency on MLIRQuantDialect. This is fine as it is unless one builds with shared libraries enabled in which case the build fails. This change adds the missing library and also does a couple of other "nice" things to improve library usage - adds the target include directories and cleans up the top-level include definition to use the variables.

We ran into this in onnx-mlir which recently took a dependency on mlir-hlo.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>
Copybara import of the project:

--
4de52638e64dc7904da6a799df475a6d46a02aad by Stella Stamenova <stilis@microsoft.com>:

[mhlo] Add missing dependency to the MhloDialect

Recent versions of mlir have moved functionality around, so now the MhloDialect has a dependency on MLIRQuantDialect. This is fine as it is unless one builds with shared libraries enabled in which case the build fails. This change adds the missing library and also does a couple of other "nice" things to improve library usage - adds the target include directories and cleans up the top-level include definition to use the variables.

We ran into this in onnx-mlir which recently took a dependency on mlir-hlo.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

Merging this change closes #56844

PiperOrigin-RevId: 462692939