Add a pass to detensorize scf.while, scf.for, scf.if.
In our lowerings, we often generate scf ops that take unit tensor
arguments. For example, a while loop that increments a counter in
each iteration typically has a tensor<i32> loop carried variable.
In the loop body (and sometimes the condition), these tensors are
always unpacked, and later a new tensor is yielded.

There are two problems with the status quo:
1. It is not desirable to use memory for these variables.
2. Bufferization can't handle these patterns efficiently, leading
   to lots of unnecessary allocs.

PiperOrigin-RevId: 472723203