PR #57530: [mhlo] fix segfault from dereferencing a non-existent value
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/57530

This patch fixes a runtime segfault caused by calling `value()` on
optional types in the predicate of `if` statements.  This patch replaces
the predicates with `has_value()` so that dereferencing the values later
is safe.

Thanks to @fortianyou for diagnosing the problem!

cc: @burmako
Copybara import of the project:

--
27d2396ad216877bf7602dd84c822441c175d214 by Ashay Rane <ashay@users.noreply.github.com>:

[mhlo] fix segfault from dereferencing a non-existent value

This patch fixes a runtime segfault caused by calling `value()` on
optional types in the predicate of `if` statements.  This patch replaces
the predicates with `has_value()` so that dereferencing the values later
is safe.

Merging this change closes #57530

PiperOrigin-RevId: 472856969