[XLA:GPU Next] Add preliminary lowering of linalg.generic to GPU warp reduce using shuffles.
This is not used anywhere yet but in a simple test. It's one of the building blocks to lower softmax to GPU. The lowering is only correct if the linalg.generic happens to be at the warp level. This is not checked anywhere, but because we currently won't ever encounter a linalg.generic with a reduction from the fusion rewrite pass, it doesn't break anything.

PiperOrigin-RevId: 477419236