[GmlSt] Generalize scalarization of thlo.concatenate.
Support a case when concatenation dim of output is statically known to be 1, but others dims can be any size. Replace `concatenate` with `tensor.insert_slice` from on of the inputs.

There is a potential room for optimization, when all dims are statically known to be 1 and `tensor.insert` can be used.

PiperOrigin-RevId: 482765141