Use `vector.insert` for scalar->vector conversion
Vectorization of GmlSt currently produces `gml_st.set_yield` to convert
from a scalar to a single-element vector. Canonicalize this by prepending `vector.insert`.

This could be thought of as a `gml_st.set_yield` canonicalization, but that
would make GmlSt dependent on the Vector dialect, thus it is done as part of
vectorization instead, once we've already opted into using the vector
dialect in our lowering.

PiperOrigin-RevId: 484495581