[THLO] Generate tiled implementation for thlo.concat based on a single operand if possible
If the requested tile is of size 1 in the concatenation dimensions, we can extract the tile from one of the operands.
In that case, there is no need to generate a tiled thlo.concat operation.
Rather, we can extract the requested tile from the unique operand directly.
This special form of tiling applied, in particular, to unit tiles, i.e. scalars.

PiperOrigin-RevId: 502960672