PR #59219: [mhlo] fix build rules for the MhloPasses target
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/59219

The MhloPasses library, which contains code from
symbolic_shape_optimization.cc, refers to
`ShapeComponentAnalysis::GetValueInfo()`, which is defined in
shape_component_analysis.cc, which is a part of the MhloAnalysis target.
However, the CMake rules for building MhloPasses do not add MhloAnalysis
as a dependency, causing undefined reference errors in downstream
targets.

As an ancillary change, this patch also removes the MhloAnalysis
dependency in the rule for building mlir-hlo-opt, since the now-fixed
dependency in MhloPasses carries over to the mlir-hlo-opt target.
Copybara import of the project:

--
55f8182265d21bea83e507e655af04096ebc6c6b by Ashay Rane <ashay@users.noreply.github.com>:

[mhlo] fix build rules for the MhloPasses target

The MhloPasses library, which contains code from
symbolic_shape_optimization.cc, refers to
`ShapeComponentAnalysis::GetValueInfo()`, which is defined in
shape_component_analysis.cc, which is a part of the MhloAnalysis target.
However, the CMake rules for building MhloPasses do not add MhloAnalysis
as a dependency, causing undefined reference errors in downstream
targets.

As an ancillary change, this patch also removes the MhloAnalysis
dependency in the rule for building mlir-hlo-opt, since the now-fixed
dependency in MhloPasses carries over to the mlir-hlo-opt target.

Merging this change closes #59219

PiperOrigin-RevId: 505985133