[mhlo] Add option for passthrough of certain MHLO features to StableHLO CustomCall encoding.
1. Add `--hlo-legalize-to-stablehlo=allow-experimental-features` flag.

By default, the encoding of MHLO ops that are public and agreed upon by XLA (currently none), will be permitted to be encoded into StableHLO custom calls. Otherwise, for ops that frontends wish to explore, the experimental flag can be set. There is still no way to encode XLA-private features, as there is no known use case to do this yet. If an op is being explored by a FE, it should be moved to the experimental category.

2. Still to-do (future): Allow attrs/types in MHLO but not StableHLO to be encoded.

PiperOrigin-RevId: 506103950