deallocation: Add a new function for alias analysis.
The current analysis critically overestimates aliasing for ops such as:

```
%c = arith.select %cond, %a, %b : memref<i32>
```

Here, we currently consider {`%c`, `%a`} and {`%c`, `%b`} aliases, which
results in {`%a`, `%b`} being considered transitive aliases. This is not
necessarily the case. The new analysis is based on graph reachability instead:
For the snippet above, we consider `%a` and `%b` to be reachable from `%c`.
When checking two values for aliasing, we look for overlap of the set of values
that are reachable from each.

PiperOrigin-RevId: 534413868