Fix --stablehlo-refine-shapes for i1 => ... conversions
This bug is a consequence of two pending items on the StableHLO side:
  * SignedIntegerType from the spec modelled as signless IntegerType from MLIR (instead of being expressed as signed integers in the first place: https://github.com/openxla/stablehlo/issues/22).
  * BooleanType from the spec modelled as IntegerType from MLIR (this has already led to some amount of confusion in the past: https://github.com/openxla/stablehlo/issues/308).

The fix addresses this by patching matchInts to correctly consider i1 to be unsigned.

As a sidenote, it would be great to eventually migrate shape computations to use the index type in MLIR. This way we won't need to spuriously constant fold unrelated integer computations like this one. It's likely going to be a lot of work on the StableHLO side though, so we probably shouldn't expect this in the near future.

PiperOrigin-RevId: 549213507