[XLA] Verify shapes for unsugared async ops
Currently at parse time we only verify the operand/resulting shapes for sugared async ops. For the unsugared case this is still being caught by the verifier, but has allowed bad HLO to slip through, specifically in the HLO <-> MHLO translater.

Async ops expect the wrapped computation's operands to be passed around wrapped in a tuple. In the case of a single operand, this diverges from old single operand async ops (e.g. all-gather-*), where we don't use a tuple. I believe this caused confusion in the translaters support of HLO async ops.

PiperOrigin-RevId: 592905794